refactor: sign up page query 분리
