+++ src/views/SignupPage/apis.ts
@@ -2,7 +2,7 @@ import instance from '@apis/instance';
 
 import { SignUpRequest } from './types';
 
-export const sendingSignUp = async (userInfo: SignUpRequest) => {
+export const sendSignUp = async (userInfo: SignUpRequest) => {
   const { email, password, passwordCheck, name, phone, season, group } = userInfo;
   const res = await instance.post('/recruiting-auth/signup', {
     email,
